ranked-web/functions/requirements.txt
firebase_functions~=0.1.0
functions-framework
google-cloud-firestore